+++ R/visualization_functions.R
@@ -12,14 +12,14 @@
 #'   path/to/PIN.sif, the file must comply with the PIN specifications. Defaults
 #'   to Biogrid.
 #'
-#' @return Depending of `gene_sets`, creates visualization of interactions of genes
+#' @return Depending on `gene_sets`, creates visualization of interactions of genes
 #' involved in the list of pathways in `result_df` and saves them in the folder
 #' "pathway_visualizations" under the current working directory.
 #'
 #'
 #' @details For `gene_sets == "KEGG"`, KEGG human pathway diagrams are created,
 #' affected nodes colored by up/down regulation status.
-#' For the rest of `gene_sets`, interactions of affected genes are determined (via a shortest-path
+#' For other `gene_sets`, interactions of affected genes are determined (via a shortest-path
 #' algorithm) and are visualized (colored by change status) using igraph.
 #'
 #'
@@ -270,6 +270,11 @@ visualize_hsa_KEGG <- function(pw_table, gene_data) {
 #' }
 #' @param plot_by_cluster boolean value indicating whether or not to group the
 #' pathways by cluster (works if "Cluster" is a column of `result_df`).
+#' @param num_bubbles number of sizes displayed in the legend `# of DEGs` (Default = 5).
+#' @param even_breaks whether or not to set even breaks for the number of sizes displayed
+#' in the legend `# of DEGs`. If `TRUE` (default), sets equal breaks and the number of
+#' displayed bubbles may be different than the number set by `num_bubbles`. If the exact
+#' number set by `num_bubbles` is required, set this argument to `FALSE`.
 #'
 #' @return a `ggplot2` object containing the bubble chart. The x-axis corresponds to
 #' fold enrichment values while the y-axis indicates the enriched pathways. Size of
@@ -283,7 +288,7 @@ visualize_hsa_KEGG <- function(pw_table, gene_data) {
 #'
 #' @examples
 #' g <- enrichment_chart(RA_output)
-enrichment_chart <- function(result_df, plot_by_cluster = FALSE) {
+enrichment_chart <- function(result_df, plot_by_cluster = FALSE, num_bubbles = 4, even_breaks = TRUE) {
 
   necessary <- c("Pathway", "Fold_Enrichment", "lowest_p", "Up_regulated", "Down_regulated")
   if (!all(necessary %in% colnames(result_df)))
@@ -309,6 +314,17 @@ enrichment_chart <- function(result_df, plot_by_cluster = FALSE) {
                           plot.title = ggplot2::element_blank())
   g <- g + ggplot2::xlab("Fold Enrichment") + ggplot2::ylab("")
   g <- g + ggplot2::labs(size = "# of DEGs", color = "-log10(lowest-p)")
+
+  ## breaks for # of DEGs
+  if (max(n) < num_bubbles) {
+    g <- g + ggplot2::scale_size_continuous(breaks = seq(0, max(n)))
+  } else {
+    if (even_breaks)
+      g <- g + ggplot2::scale_size_continuous(breaks = base::seq(0, max(n), round(max(n) / (num_bubbles + 1))))
+    else
+      g <- g + ggplot2::scale_size_continuous(breaks = base::round(base::seq(0, max(n), length.out = num_bubbles + 1)))
+  }
+
   g <- g + ggplot2::scale_color_continuous(low = "#f5efef", high = "red")
 
   if (plot_by_cluster & "Cluster" %in% colnames(result_df)) {